# Optimize new DateTime(const, const, const)