@EgorBo 2022-02-01
Closing as it's unlikely the hack worth it or people use this pattern on hot paths, but it should be a good challenge for jit to do everything under the hood.

Another note that my single line change in `morph.cpp` seems to trigger some diffs: `Total bytes of delta: -1496 (-0.00 % of base)`